data-engineering-zoomcamp/2026/Homework5/pipeline/assets/ingestion/requirements.txt
pandas
requests
pyarrow
python-dateutil